refactor: move AppStart logic from OnboardingUtils to AppStartUtils
